Rename xterm.js to Terminal.js
xterm.js simply exports Terminal so it makes sense to name it Terminal.

Fixes #373
