+++ src/Terminal.js
