+++ src/test/addons/test.js
@@ -1,5 +1,5 @@
 var assert = require('chai').assert;
-var Terminal = require('../../xterm');
+var Terminal = require('../../Terminal');
 
 describe('xterm.js addons', function() {
   it('should load addons with Terminal.loadAddon', function () {
